fix: Fixing HTML code errors
Fix the link of the "Shop now" button to work correctly.
Add appropriate headers to old and new product articles.
Change the 'section' element with class 'two-discounts' to a 'div' for better semantics.
Replace the 'div' with class 'cont-footer' with classes on the 'li' elements.
Resolve a bug that affects the description on mobile devices.